+++ Sources/DTCollectionViewManager/DTCollectionViewManager+Delegate.swift
@@ -31,7 +31,21 @@ import UIKit
 import TVUIKit
 #endif
 
-/// Extension for UICollectionViewDelegate events
+/// Extension for UICollectionViewDelegate events. Please note that cell / view methods in this extension are soft-deprecated, and it's recommended to migrate to methods extending `CellViewModelMappingProtocolGeneric` and `SupplementaryViewModelMappingProtocolGeneric`:
+///
+/// Deprecated:
+/// ```swift
+///     manager.register(PostCell.self)
+///     manager.didSelect(PostCell.self) { postCell, post, indexPath in }
+/// ```
+/// Recommended:
+/// ```swift
+///     manager.register(PostCell.self) { mapping in
+///         mapping.didSelect { postCell, post, indexPath in }
+///     }
+/// ```
+/// While previously main benefits for second syntax were mostly syntactic, now with support for SwiftUI it will be hard to actually specialize hosting cells, so only second syntax will work for all kinds of cells, and first syntax can only work for non-SwiftUI cells.
+/// New delegate methods for UICollectionView (starting with iOS 16 / tvO 16 SDK) will be added only as extension to mapping protocols, not DTCollectionViewManager itself.
 public extension DTCollectionViewManager {
     /// Registers `closure` to be executed, when `UICollectionViewDelegate.collectionView(_:didSelectItemAt:)` method is called for `cellClass`.
     func didSelect<Cell:ModelTransfer>(_ cellClass:  Cell.Type, _ closure: @escaping (Cell, Cell.ModelType, IndexPath) -> Void) where Cell:UICollectionViewCell
@@ -493,6 +507,18 @@ public extension CellViewModelMappingProtocolGeneric where Cell: UICollectionVie
     {
         reactions.append(FiveArgumentsEventReaction(Cell.self, modelType: Model.self, argumentOne: IndexPath.self, argumentTwo: IndexPath.self, signature: EventMethodSignature.targetIndexPathForMoveOfItemFromOriginalIndexPath.rawValue, closure))
     }
+    
+#if swift(>=5.7) || (os(macOS) && swift(>=5.7.1))  // Xcode 14.0 AND macCatalyst on Xcode 14.1
+    @available(iOS 16, tvOS 16, *)
+    func canPerformPrimaryAction(_ closure: @escaping (Cell, Model, IndexPath) -> Bool) {
+        reactions.append(EventReaction(viewType: Cell.self, modelType: Model.self, signature: EventMethodSignature.canPerformActionForItemAtIndexPath.rawValue, closure))
+    }
+    
+    @available(iOS 16, tvOS 16, *)
+    func performPrimaryAction(_ closure: @escaping (Cell, Model, IndexPath) -> Void) {
+        reactions.append(EventReaction(viewType: Cell.self, modelType: Model.self, signature: EventMethodSignature.performPrimaryActionForItemAtIndexPath.rawValue, closure))
+    }
+#endif
     
     @available(iOS 14, tvOS 14, *)
     /// Registers `closure` to be executed, when `UICollectionViewDelegate.collectionView(_:canEditItemAt:)` method is called.